# api/controllers/gamesystem.controller.js
'use strict';
var controllerHelper = require('../helpers/controller.helper');

const {Gamesystems} = require('../models');

// Module Name
const MODULE_NAME = '[GameSystem Controller]';

// Error Messages
const GS_CT_ERR_GAMESYSTEM_NOT_FOUND = 'Game system not found';

// Success Messages
const GS_CT_DELETED_SUCCESSFULLY = 'Game system deleted successfully';

function getGameSystems(req, res) {
    try {

        console.log("fetching...");
        console.log(Gamesystems);

        Gamesystems.findAll(
            {
                //include: [{model: orderstatus}],
                //include: [{all: true, nested: true}]
            }).then((consoles) =>
            {
                console.log(consoles);
                res.status(200).send(consoles);
                //utils.writeJson(res, consoles);
            }, (error) => { res.status(500).send(error); });

    } catch (error) {
        controllerHelper.handleErrorResponse(MODULE_NAME, getGameSystems.name, error, res);
    }
}

function createGameSystem(req, res) {
    res.setHeader('Access-Control-Allow-Origin', '*');
    res.setHeader('Access-Control-Allow-Methods', 'GET, POST, OPTIONS, PUT, PATCH, DELETE'); // If needed
    res.setHeader('Access-Control-Allow-Headers', 'X-Requested-With, Content-Type'); // If needed
    res.header('Access-Control-Allow-Headers', 'Content-Type');

    try {
        console.log("Insert...");
        const parameters = req.body;
        console.log("params... " + parameters);

        return Gamesystems.create(
            {
                name: parameters.name,
                description: parameters.description,
            },
            {
                /*include: [ {model: order_detail, as: 'orderdetail'} ]*/
            }).then((parameters) =>
            {
                res.status(201).send(parameters); //mygamesystem
            }).catch((error) => res.status(400).send(error));

    } catch (error) {
        console.log("Was an error");
        controllerHelper.handleErrorResponse(MODULE_NAME, createGameSystem.name, error, res);
    }

}

function getGameSystemById(req, res) {
    //console.log("operadores.controller getOperadorById");
    try {
        console.log(req.swagger.params.id.value);
        var id = req.swagger.params.id.value;
        console.log("gamesystem by id..." + id);
        //console.log(gamesystems);

        Gamesystems.findByPk(id).then(mygamesystem => {
            console.log(mygamesystem);
            res.status(200).send(mygamesystem);
        })
    } catch (error) {
        console.log("Was an error");
        controllerHelper.handleErrorResponse(MODULE_NAME, getGameSystemById.name, error, res);
    }
}

function deleteGameSystem(req, res) {

    res.setHeader('Access-Control-Allow-Origin', '*');

    res.setHeader('Access-Control-Allow-Methods', 'GET, POST, OPTIONS, PUT, PATCH, DELETE'); // If needed

    res.setHeader('Access-Control-Allow-Headers', 'X-Requested-With,contenttype'); // If needed

    res.header('Access-Control-Allow-Headers', 'Content-Type');

    console.log(req.swagger.params.id.value);

    var id = req.swagger.params.id.value;

    Gamesystems

        .findByPk(id)

        .then(mygamesystem => {

            console.log("Result of findById: " + mygamesystem);

            if (!mygamesystem) {

                res.status(200).send({"success": 0, "description": "not found !"});

            } else {

                return mygamesystem

                    .destroy()

                    .then(() => res.status(200).send({"success": 1, "description": "deleted!"}))

                    .catch(error => res.status(403).send({"success": 0, "description": "error !"}))

            }

        })

        .catch(error => {

            console.log("There was an error: " + error);

        });

}

function updateGameSystem(req, res) {

    res.setHeader('Access-Control-Allow-Origin', '*');

    res.setHeader('Access-Control-Allow-Methods', 'GET, POST, OPTIONS, PUT, PATCH, DELETE'); // If needed

    res.setHeader('Access-Control-Allow-Headers', 'X-Requested-With,contenttype'); // If needed

    res.header('Access-Control-Allow-Headers', 'Content-Type');

//console.log("operadores.controller getOperadorById");

    try {

        var id = req.swagger.params.id.value;

        console.log("params : " + id);

        var myupdategamesystem = req.body;

        console.log("update gamesystems ... " + myupdategamesystem.name + " " + myupdategamesystem.description);

        Gamesystems.findByPk(id)

            .then(mygamesystem => {

                console.log("Result of findById: " + mygamesystem);

                if (!mygamesystem) {

                    res.status(401).send(({}));

                }

                return mygamesystem

                    .update({

                        name: myupdategamesystem.name,

                        description: myupdategamesystem.description

                    })

                    .then(() => res.status(200).send(mygamesystem))

                    .catch(error => res.status(403).send(mygamesystem));

            })

            .catch(error => {

                console.log("There was an error: " + error);

//resolve(error);

            });

    } catch (error) {

        console.log("Was an error");

        controllerHelper.handleErrorResponse(MODULE_NAME, updateGameSystem.name, error, res);

    }

}


module.exports =
{
    getGameSystems,
    getGameSystemById,
    createGameSystem,
    updateGameSystem,
    deleteGameSystem,
    GS_CT_ERR_GAMESYSTEM_NOT_FOUND,
    GS_CT_DELETED_SUCCESSFULLY,
    MODULE_NAME
};